refactor: Review changes
